backend_database/pyproject.toml
[project]
name = "backend-database"
version = "0.1.0"
description = "Django and Postgres internal API layer"
authors = [
    {name = "Let's Do Digital"}
]
requires-python = ">=3.13,<3.14"
dependencies = [
    "django (>=5.2,<6.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
